@@ -4,14 +4,14 @@
     <parent>
         <groupId>org.cardanofoundation</groupId>
         <artifactId>hydra-java</artifactId>
-        <version>0.0.8</version>
+        <version>0.0.9-SNAPSHOT</version>
     </parent>
 
     <artifactId>hydra-java-cardano-client-lib-adapter</artifactId>
 
     <properties>
-        <maven.compiler.source>11</maven.compiler.source>
-        <maven.compiler.target>11</maven.compiler.target>
+        <maven.compiler.source>17</maven.compiler.source>
+        <maven.compiler.target>17</maven.compiler.target>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
     </properties>
 
@@ -20,7 +20,7 @@
         <dependency>
             <groupId>org.cardanofoundation</groupId>
             <artifactId>hydra-java-core</artifactId>
-            <version>0.0.7</version>
+            <version>0.0.9-SNAPSHOT</version>
         </dependency>
 
         <dependency>
@@ -33,6 +33,7 @@
             <groupId>com.bloxbean.cardano</groupId>
             <artifactId>cardano-client-transaction-spec</artifactId>
         </dependency>
+
         <dependency>
             <groupId>com.bloxbean.cardano</groupId>
             <artifactId>cardano-client-function</artifactId>
