@@ -12,7 +12,8 @@
 
     <properties>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-        <java.version>11</java.version>
+
+        <java.version>17</java.version>
 
         <maven-compiler-plugin.version>3.10.1</maven-compiler-plugin.version>
         <slf4j.version>2.0.7</slf4j.version>
