@@ -39,8 +39,9 @@ public class HydraWSClientIntegrationTest2 {
             var aliceState = new AtomicReference<HydraState>();
             var bobState = new AtomicReference<HydraState>();
 
-            var aliceHydraWSClient = new HydraWSClient(HydraClientOptions.builder(HydraDevNetwork.getHydraApiUrl(hydraDevNetwork.getAliceHydraContainer()))
+            var aliceHydraWSClient = new HydraWSClient(HydraClientOptions.builder(HydraDevNetwork.getHydraApiWebSocketUrl(hydraDevNetwork.getAliceHydraContainer()))
                     .utxoStore(new InMemoryUTxOStore())
+                    .snapshotUtxo(true)
                     .build());
             SLF4JHydraLogger aliceHydraLogger = SLF4JHydraLogger.of(log, "alice");
             aliceHydraWSClient.addHydraQueryEventListener(aliceHydraLogger);
@@ -54,8 +55,9 @@ public class HydraWSClientIntegrationTest2 {
                 }
             });
 
-            var bobHydraWSClient = new HydraWSClient(HydraClientOptions.builder(HydraDevNetwork.getHydraApiUrl(hydraDevNetwork.getBobHydraContainer()))
+            var bobHydraWSClient = new HydraWSClient(HydraClientOptions.builder(HydraDevNetwork.getHydraApiWebSocketUrl(hydraDevNetwork.getBobHydraContainer()))
                     .utxoStore(new InMemoryUTxOStore())
+                    .snapshotUtxo(true)
                     .build());
 
             SLF4JHydraLogger bobHydraLogger = SLF4JHydraLogger.of(log, "bob");
