@@ -6,7 +6,7 @@ import com.bloxbean.cardano.client.exception.CborSerializationException;
 import com.bloxbean.cardano.client.function.Output;
 import com.bloxbean.cardano.client.function.TxBuilderContext;
 import lombok.val;
-import org.cardanofoundation.hydra.cardano.client.lib.HydraOperator;
+import org.cardanofoundation.hydra.cardano.client.lib.wallet.CardanoOperator;
 
 import static com.bloxbean.cardano.client.common.ADAConversionUtil.adaToLovelace;
 import static com.bloxbean.cardano.client.common.CardanoConstants.LOVELACE;
@@ -16,6 +16,7 @@ import static com.bloxbean.cardano.client.function.helper.InputBuilders.createFr
 public class HydraTransactionGenerator {
 
     private final UtxoSupplier utxoSupplier;
+
     private final ProtocolParamsSupplier protocolParamsSupplier;
 
     public HydraTransactionGenerator(UtxoSupplier utxoSupplier,
@@ -24,7 +25,9 @@ public class HydraTransactionGenerator {
         this.protocolParamsSupplier = protocolParamsSupplier;
     }
 
-    public byte[] simpleTransfer(HydraOperator sender, HydraOperator receiver, int adaAmount) throws CborSerializationException {
+    public byte[] simpleTransfer(CardanoOperator sender,
+                                 CardanoOperator receiver,
+                                 int adaAmount) throws CborSerializationException {
         var senderAddress = sender.getAddress();
         var receiverAddress = receiver.getAddress();
 
