@@ -10,7 +10,9 @@ public class Response {
     protected final int seq;
     protected final boolean isFailure;
 
-    protected Response(Tag tag, int seq, boolean isFailure) {
+    protected Response(Tag tag,
+                       int seq,
+                       boolean isFailure) {
         this.tag = tag;
         this.seq = seq;
         this.isFailure = isFailure;
