@@ -1,16 +1,10 @@
 package org.cardanofoundation.hydra.reactor;
 
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.ToString;
 import org.cardanofoundation.hydra.core.model.HydraState;
 
-@AllArgsConstructor
-@Getter
-@ToString
-public class BiHydraState {
-
-    private HydraState oldState;
-    private HydraState newState;
-
+/**
+ * Represents a pair of Hydra states, typically representing the transition from an old state to a new state.
+ * This class is annotated with Lombok's annotations for generating constructors, getters, and toString methods.
+ */
+public record BiHydraState(HydraState oldState, HydraState newState) {
 }
